Added mouselook
The ship and camera now rotate based on mouse movement. To get it to
work properly, I had to disable a script that someone put on the Main
Camera. Sorry about that.

For some reason, though, all the controls are now inverted. W moves
back, S moves forward, A moves right, and D moves left. I should get a
chance to look at it again before the meeting today.